Linter error resolutions.

+++ Libraries/CMSIS/Device/Maxim/MAX32650/Source/system_max32650.c
@@ -112,7 +112,7 @@ __weak void SystemInit(void)
     ChipRevision = MXC_SYS_GetRev();
     /* Configure the interrupt controller to use the application vector table in */
     /* the application space */
-    SCB->VTOR = (unsigned long)__isr_vector;
+    SCB->VTOR = (uint32_t)__isr_vector;
 
     /* MAX3265x ROM turns off interrupts, which is not the same as the reset state. */
     __enable_irq();
