+++ Libraries/CMSIS/Device/Maxim/MAX32675/Source/system_max32675.c
@@ -122,11 +122,11 @@ __weak void SystemInit(void)
     /* the application space */
 #if defined(__CC_ARM) || defined(__GNUC__)
     /* IAR sets the VTOR pointer incorrectly and causes stack corruption */
-    SCB->VTOR = (unsigned long)__isr_vector;
+    SCB->VTOR = (uint32_t)__isr_vector;
 #endif /* __CC_ARM || __GNUC__ */
 
 #if defined __ICCARM__
-    SCB->VTOR = (unsigned long)__vector_table;
+    SCB->VTOR = (uint32_t)__vector_table;
 #endif
 
     /* Make sure interrupts are enabled. */
